Remove `.boxed()` requirement for `context_extractor` filter
